day10/langchain_chat_app/requirements.txt
langchain-openai
langchain
python-dotenv
flask
flask-cors
requests
pytest
